exercises/part4/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
werkzeug==3.0.1
